Add files via upload
The is simple project it is one of One of the popular projects in terms of understanding the object & class structure in Javascript is ToDo List. It is not hard to make up this project, yet it requires attention in certain points.
I share it with you guys and how did i build it.